Renaming files to match repo convention
